Code review fix
